refactor(guard): split daemon hook transport
