+++ ci/test-suite-ratchet-baseline.json
@@ -1,5 +1,5 @@
 {
   "maximum_collected_cases": 11870,
-  "maximum_test_source_lines": 278043,
+  "maximum_test_source_lines": 278085,
   "schema_version": 1
 }
