+++ src/codex_plugin_scanner/guard/adapters/codex_daemon_hook_bridge.py
@@ -2,19 +2,14 @@
 
 from __future__ import annotations
 
-import hashlib
-import hmac
 import http.client
 import json
-import os
 import secrets
-import stat
 import sys
 import time
 import urllib.error
 from collections.abc import Callable, Mapping, Sequence
 from pathlib import Path
-from urllib.parse import urlparse
 
 if __package__:
     from ..codex_hook_bridge_runtime import BridgeConfig
@@ -24,10 +19,19 @@ if __package__:
     from ..codex_hook_bridge_runtime import trusted_hook_launch as _trusted_hook_launch
     from ..codex_hook_launch_runtime import isolated_hook_environment as _isolated_hook_environment
     from ..codex_hook_launch_runtime import run_isolated_hook_process as _run_isolated_hook_process
+    from .codex_daemon_hook_auth import _DaemonResponseError
+    from .codex_daemon_hook_transport import _daemon_response_once, _DaemonGenerationChangedError
 else:  # pragma: no cover - exercised by subprocess integration tests
     _package_root = str(Path(__file__).resolve().parents[3])
     if _package_root not in sys.path:
         sys.path.insert(0, _package_root)
+    from codex_plugin_scanner.guard.adapters.codex_daemon_hook_auth import (
+        _DaemonResponseError,
+    )
+    from codex_plugin_scanner.guard.adapters.codex_daemon_hook_transport import (
+        _daemon_response_once,
+        _DaemonGenerationChangedError,
+    )
     from codex_plugin_scanner.guard.codex_hook_bridge_runtime import (
         BridgeConfig,
     )
@@ -50,12 +54,9 @@ else:  # pragma: no cover - exercised by subprocess integration tests
         run_isolated_hook_process as _run_isolated_hook_process,
     )
 
-_LOOPBACK_HOSTS = frozenset({"127.0.0.1", "localhost", "::1"})
 _HOOK_TIMEOUT_GRACE_SECONDS = 2
 _DAEMON_START_TIMEOUT_SECONDS = 8
 _DISCOVERY_PROTOCOL_VERSION = 1
-_DISCOVERY_CHALLENGE_TTL_SECONDS = 5
-_MAX_DAEMON_RESPONSE_BYTES = 1_000_000
 _MAX_HOOK_INPUT_BYTES = 1_000_000
 _FAIL_CLOSED_REASON = "HOL Guard could not authenticate the local daemon. Run `hol-guard daemon repair`, then retry."
 _LAUNCH_INTEGRITY_REASON = (
@@ -68,45 +69,12 @@ _OVERLOAD_REASON = (
 )
 
 
-class _DaemonResponseError(ValueError):
-    def __init__(self, status: int, detail: str, *, authenticated: bool) -> None:
-        super().__init__(f"daemon returned HTTP {status}")
-        self.status = status
-        self.detail = detail
-        self.authenticated = authenticated
-
-
-class _DaemonGenerationChangedError(ValueError):
-    """The authenticated daemon generation rotated during one hook request."""
-
-
-def _assert_loopback_http_url(url: str) -> None:
-    parsed = urlparse(url)
-    if parsed.scheme != "http":
-        raise ValueError(f"daemon URL must use http, not {parsed.scheme!r}")
-    if parsed.username is not None or parsed.password is not None:
-        raise ValueError("daemon URL must not contain credentials")
-    host = (parsed.hostname or "").lower()
-    if host not in _LOOPBACK_HOSTS:
-        raise ValueError(f"daemon URL must target loopback, not {host!r}")
-    if parsed.port is None:
-        raise ValueError("daemon URL must include an explicit port")
-
-
 def _json_object(text: str) -> dict[str, object] | None:
     try:
         payload = json.loads(text)
     except json.JSONDecodeError:
         return None
     return payload if isinstance(payload, dict) else None
-
-
-def _with_remaining_hint(data: str, deadline: float) -> str:
-    payload = _json_object(data)
-    if payload is None:
-        return data
-    payload["guard_remaining_ms"] = min(60_000, max(1, int((deadline - time.monotonic()) * 1000)))
-    return json.dumps(payload, ensure_ascii=True, separators=(",", ":"))
 
 
 def _transient_overload(error: BaseException) -> tuple[int, int] | None:
@@ -140,190 +108,6 @@ def _retry_transient_overload(
         return None
     time.sleep(jitter_ms / 1000)
     return request()
-
-
-def _private_file_text(path: Path, *, label: str) -> str:
-    try:
-        parent_metadata = path.parent.lstat()
-        metadata = path.lstat()
-    except OSError as error:
-        raise ValueError(f"{label} is unavailable") from error
-    if not stat.S_ISDIR(parent_metadata.st_mode):
-        raise ValueError("Guard home is not a directory")
-    if stat.S_ISLNK(metadata.st_mode) or not stat.S_ISREG(metadata.st_mode):
-        raise ValueError(f"{label} must be a regular file")
-    if os.name != "nt":
-        if parent_metadata.st_uid != os.getuid() or metadata.st_uid != os.getuid():
-            raise ValueError(f"{label} ownership does not match the current user")
-        if stat.S_IMODE(parent_metadata.st_mode) & 0o077 or stat.S_IMODE(metadata.st_mode) & 0o077:
-            raise ValueError(f"{label} permissions are not owner-only")
-    try:
-        return path.read_text(encoding="utf-8").strip()
-    except OSError as error:
-        raise ValueError(f"{label} is unreadable") from error
-
-
-def _canonical_discovery_payload(payload: dict[str, object]) -> bytes:
-    return json.dumps(payload, sort_keys=True, separators=(",", ":"), ensure_ascii=True).encode("utf-8")
-
-
-def _sign_discovery_payload(discovery_key: str, payload: dict[str, object]) -> str:
-    try:
-        key = bytes.fromhex(discovery_key)
-    except ValueError as error:
-        raise ValueError("daemon discovery key is malformed") from error
-    if len(key) != 32:
-        raise ValueError("daemon discovery key is malformed")
-    return hmac.new(key, _canonical_discovery_payload(payload), hashlib.sha256).hexdigest()
-
-
-def _authenticated_state(state_path: str | Path) -> tuple[dict[str, object], str]:
-    path = Path(state_path)
-    discovery_key = _private_file_text(path.parent / "daemon-discovery-key", label="daemon discovery key")
-    try:
-        payload = json.loads(_private_file_text(path, label="daemon state"))
-    except json.JSONDecodeError as error:
-        raise ValueError("daemon state is malformed") from error
-    if not isinstance(payload, dict):
-        raise ValueError("daemon state must be a JSON object")
-    signature = payload.get("state_signature")
-    unsigned = {key: value for key, value in payload.items() if key != "state_signature"}
-    try:
-        expected_key_id = hashlib.sha256(bytes.fromhex(discovery_key)).hexdigest()
-    except ValueError as error:
-        raise ValueError("daemon discovery key is malformed") from error
-    if (
-        not isinstance(signature, str)
-        or unsigned.get("discovery_protocol_version") != _DISCOVERY_PROTOCOL_VERSION
-        or unsigned.get("discovery_key_id") != expected_key_id
-        or not secrets.compare_digest(signature, _sign_discovery_payload(discovery_key, unsigned))
-    ):
-        raise ValueError("daemon state authentication failed")
-    host = unsigned.get("host")
-    port = unsigned.get("port")
-    pid = unsigned.get("pid")
-    state_id = unsigned.get("state_id")
-    started_at = unsigned.get("started_at")
-    guard_home = unsigned.get("guard_home")
-    auth_token_id = unsigned.get("auth_token_id")
-    if (
-        not isinstance(host, str)
-        or host.lower() not in _LOOPBACK_HOSTS
-        or not isinstance(port, int)
-        or not 0 < port <= 65535
-        or not isinstance(pid, int)
-        or pid <= 0
-        or not isinstance(state_id, str)
-        or not state_id
-        or not isinstance(started_at, str)
-        or not started_at
-        or not isinstance(guard_home, str)
-        or not isinstance(auth_token_id, str)
-    ):
-        raise ValueError("daemon state identity is incomplete")
-    try:
-        expected_guard_home = str(path.parent.resolve())
-        state_guard_home = str(Path(guard_home).resolve())
-    except OSError as error:
-        raise ValueError("daemon state Guard home is invalid") from error
-    if state_guard_home != expected_guard_home:
-        raise ValueError("daemon state belongs to a different Guard home")
-    return payload, discovery_key
-
-
-def _daemon_url(state_path: str | Path) -> str:
-    payload, _discovery_key = _authenticated_state(state_path)
-    host = str(payload["host"])
-    port = payload.get("port")
-    rendered_host = f"[{host}]" if ":" in host else host
-    return f"http://{rendered_host}:{port}"
-
-
-def _daemon_auth_token(state_path: str | Path, state: Mapping[str, object]) -> str:
-    path = Path(state_path)
-    token = _private_file_text(path.parent / "daemon-auth-token", label="daemon auth token")
-    expected_token_id = state.get("auth_token_id")
-    actual_token_id = hashlib.sha256(token.encode("utf-8")).hexdigest()
-    if (
-        not token
-        or not isinstance(expected_token_id, str)
-        or not secrets.compare_digest(actual_token_id, expected_token_id)
-    ):
-        raise ValueError("daemon auth token does not match authenticated state")
-    return token
-
-
-def _http_json_response(
-    response: http.client.HTTPResponse,
-    *,
-    label: str,
-    connection: http.client.HTTPConnection,
-    deadline: float,
-    authenticated: bool,
-) -> dict[str, object]:
-    body = bytearray()
-    while len(body) <= _MAX_DAEMON_RESPONSE_BYTES:
-        remaining = _remaining_seconds(deadline)
-        if remaining < _MINIMUM_OPERATION_SECONDS:
-            raise TimeoutError(f"{label} exceeded the hook deadline")
-        if connection.sock is not None:
-            connection.sock.settimeout(remaining)
-        chunk = response.read1(min(64 * 1024, _MAX_DAEMON_RESPONSE_BYTES + 1 - len(body)))
-        if not chunk:
-            break
-        body.extend(chunk)
-    if len(body) > _MAX_DAEMON_RESPONSE_BYTES:
-        raise ValueError(f"{label} response is too large")
-    if response.status != 200:
-        raise _DaemonResponseError(
-            response.status,
-            body.decode("utf-8", errors="replace").strip(),
-            authenticated=authenticated,
-        )
-    payload = _json_object(bytes(body).decode("utf-8", errors="replace").strip())
-    if payload is None:
-        raise ValueError(f"{label} returned malformed JSON")
-    return payload
-
-
-def _verify_challenge_response(
-    response: dict[str, object],
-    *,
-    state: Mapping[str, object],
-    discovery_key: str,
-    nonce: str,
-    hook_event: str,
-) -> str:
-    proof = response.get("proof")
-    unsigned = {key: value for key, value in response.items() if key != "proof"}
-    expected_fields = {
-        "protocol_version": _DISCOVERY_PROTOCOL_VERSION,
-        "nonce": nonce,
-        "state_id": state.get("state_id"),
-        "host": state.get("host"),
-        "port": state.get("port"),
-        "pid": state.get("pid"),
-        "started_at": state.get("started_at"),
-        "guard_home": state.get("guard_home"),
-        "hook_event": hook_event,
-    }
-    if any(unsigned.get(key) != value for key, value in expected_fields.items()):
-        raise ValueError("daemon identity challenge did not match authenticated state")
-    issued_at_ms = unsigned.get("issued_at_ms")
-    expires_at_ms = unsigned.get("expires_at_ms")
-    now_ms = int(time.time() * 1000)
-    if (
-        not isinstance(issued_at_ms, int)
-        or not isinstance(expires_at_ms, int)
-        or issued_at_ms > now_ms + 1000
-        or expires_at_ms < now_ms
-        or expires_at_ms - issued_at_ms > _DISCOVERY_CHALLENGE_TTL_SECONDS * 1000
-    ):
-        raise ValueError("daemon identity challenge expired")
-    expected_proof = _sign_discovery_payload(discovery_key, unsigned)
-    if not isinstance(proof, str) or not secrets.compare_digest(proof, expected_proof):
-        raise ValueError("daemon identity challenge authentication failed")
-    return proof
 
 
 def _event_name(data: str) -> str:
@@ -418,110 +202,6 @@ def _run_local_fallback(
     if not result.stdout.strip():
         return {}
     return _json_object(result.stdout.strip())
-
-
-def _daemon_generation_identity(state: Mapping[str, object]) -> tuple[object, ...]:
-    return tuple(state.get(field) for field in ("state_id", "auth_token_id", "host", "port", "pid", "started_at"))
-
-
-def _daemon_response_once(
-    *,
-    state_path: str | Path,
-    query: str,
-    data: str,
-    timeout_seconds: float,
-) -> dict[str, object] | None:
-    deadline = time.monotonic() + max(0.0, timeout_seconds)
-    state, discovery_key = _authenticated_state(state_path)
-    host = str(state["host"])
-    port_value = state["port"]
-    if not isinstance(port_value, int):
-        raise ValueError("daemon state port is invalid")
-    port = port_value
-    rendered_host = f"[{host}]" if ":" in host else host
-    endpoint = f"http://{rendered_host}:{port}/v1/hooks/codex?{query}"
-    _assert_loopback_http_url(endpoint)
-    hook_event = _event_name(data)
-    nonce = secrets.token_hex(32)
-    connection = http.client.HTTPConnection(host, port, timeout=timeout_seconds)
-    try:
-        challenge_body = json.dumps(
-            {
-                "protocol_version": _DISCOVERY_PROTOCOL_VERSION,
-                "nonce": nonce,
-                "state_id": state["state_id"],
-                "hook_event": hook_event,
-            },
-            separators=(",", ":"),
-        )
-        connection.request(
-            "POST",
-            "/v1/daemon/identity-challenge",
-            body=challenge_body.encode("utf-8"),
-            headers={"Content-Type": "application/json", "Connection": "keep-alive"},
-        )
-        challenge = _http_json_response(
-            connection.getresponse(),
-            label="daemon identity challenge",
-            connection=connection,
-            deadline=deadline,
-            authenticated=False,
-        )
-        proof = _verify_challenge_response(
-            challenge,
-            state=state,
-            discovery_key=discovery_key,
-            nonce=nonce,
-            hook_event=hook_event,
-        )
-        current_state, current_key = _authenticated_state(state_path)
-        if current_state != state or not secrets.compare_digest(current_key, discovery_key):
-            raise _DaemonGenerationChangedError("daemon state changed during identity verification")
-        try:
-            auth_token = _daemon_auth_token(state_path, state)
-        except ValueError as error:
-            current_state, current_key = _authenticated_state(state_path)
-            if current_state != state or not secrets.compare_digest(current_key, discovery_key):
-                raise _DaemonGenerationChangedError("daemon state changed before token authentication") from error
-            raise
-        remaining = _remaining_seconds(deadline)
-        if remaining < _MINIMUM_OPERATION_SECONDS:
-            raise TimeoutError("daemon identity challenge exhausted the hook deadline")
-        connection.timeout = remaining
-        if connection.sock is not None:
-            connection.sock.settimeout(remaining)
-        hook_path = f"/v1/hooks/codex?{query}"
-        hinted_data = _with_remaining_hint(data, deadline)
-        connection.request(
-            "POST",
-            hook_path,
-            body=hinted_data.encode("utf-8"),
-            headers={
-                "Content-Type": "application/json",
-                "Connection": "close",
-                "X-Guard-Token": auth_token,
-                "X-Guard-Daemon-Nonce": nonce,
-                "X-Guard-Daemon-Proof": proof,
-            },
-        )
-        try:
-            return _http_json_response(
-                connection.getresponse(),
-                label="daemon hook",
-                connection=connection,
-                deadline=deadline,
-                authenticated=True,
-            )
-        except _DaemonResponseError as error:
-            if error.status in {401, 403}:
-                current_state, current_key = _authenticated_state(state_path)
-                if _daemon_generation_identity(current_state) != _daemon_generation_identity(
-                    state
-                ) or not secrets.compare_digest(current_key, discovery_key):
-                    raise _DaemonGenerationChangedError("daemon state changed before hook authentication") from error
-            raise
-    finally:
-        connection.close()
 
 
 def _daemon_response(
